Refactoring enable / disable element scripts
There are now two scripts that can be equipped to the RevealElement object. ElementDiableAndEnable has both onTrigger enter and exit functions and thus forces the user to remain on that element if they want invisible objects to appear.
ElementDisableOrEnable has a boolean the programmer can set - if deselected, the chosen object will disappear when the element is entered, if selected it will appear.